% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{calc_dist}
\alias{calc_dist}
\title{calc_dist}
\usage{
calc_dist(file)
}
\arguments{
\item{file}{Mouse track containing x any y coordinates#'}
}
\value{
mouse track including distance#'
}
\description{
calc_dist
}
\details{
Calculate euclidean distance between track points
}
\examples{
calc_dist(mouse_track)

}
