% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmouse-package.R
\docType{package}
\name{devis}
\alias{devis}
\alias{devis-package}
\title{devis}
\description{
Tracks and visualizes mouse movement data. blablabla
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marcosci/devis}
  \item Report bugs at \url{https://github.com/marcosci/devis/issues}
}

}
\author{
\strong{Maintainer}: Marco Sciaini \email{marco.sciaini@indblik.io}

Authors:
\itemize{
  \item Cedric Scherer
  \item Stefan Reifenberg
}

}
