% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviz.R
\docType{package}
\name{deviz}
\alias{deviz}
\alias{deviz-package}
\title{deviz}
\description{
Tracks and visualizes mouse movement data. blablabla
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marcosci/deviz}
  \item Report bugs at \url{https://github.com/marcosci/deviz/issues}
}

}
\author{
\strong{Maintainer}: Marco Sciaini \email{marco.sciaini@indblik.io}

Authors:
\itemize{
  \item Cedric Scherer
  \item Stefan Reifenberg
}

}
