% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmouse-package.R
\docType{package}
\name{ggmouse}
\alias{ggmouse}
\alias{ggmouse-package}
\title{ggmouse}
\description{
Tracks and visualizes mouse movement data. blablabla
}
\author{
\strong{Maintainer}: Marco Sciaini \email{marco.sciaini@indblik.io}

Authors:
\itemize{
  \item Cedric Scherer
  \item Stefan Reifenberg
}

}
