% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mouse.R
\name{import_mouse}
\alias{import_mouse}
\title{import_mouse}
\usage{
import_mouse(file)
}
\arguments{
\item{file}{Path and filename, file format must currently be ".txt"}
}
\description{
import_mouse
}
\examples{
\dontrun{
import_mouse("data/test.txt")
}
}
