% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{key_df}
\alias{key_df}
\title{Keyboard stroke data}
\format{
A data frame with 14,251 rows and 8 variables:
\describe{
  \item{key}{Pressed key}
  \item{date}{Date and time of event}
}
}
\usage{
key_df
}
\description{
A dataset containing the x and y coordinates of a 20 minute
session of one of the authors on a 3.840 x 2.160 Pixel
screen.
}
\keyword{datasets}
