% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_track}
\alias{mouse_track}
\title{Computer mouse movement data}
\format{
A data frame with 3703498 rows and 2 variables:
\describe{
  \item{x}{Coordinate on x-axis of the screen}
  \item{y}{Coordinate on y-axis of the screen}
}
}
\usage{
mouse_track
}
\description{
A dataset containing the x and y coordinates of an
8 hour working day of one of the authors on a 3.840 x 2.160 Pixel
screen.
}
\keyword{datasets}
