% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_keyboard.R
\name{track_keyboard}
\alias{track_keyboard}
\title{track_keyboard}
\usage{
track_keyboard(time, as_job = FALSE)
}
\arguments{
\item{time}{Time format, "99h99m99s"}

\item{as_job}{Should the function be started as RStudio Job? (logical)}
}
\description{
track_keyboard
}
\details{
track_keyboard is a function that is best run as a background job, since it blocks
the R shell. See our README for a walkthrough of this.

If you run
}
\examples{
\dontrun{
track_mouse("0h10m00s", "mouse_position.txt")
}
}
