% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_mouse.R
\name{visualize_mouse}
\alias{visualize_mouse}
\title{track_mouse}
\usage{
visualize_mouse(track, type = "tinygrid")
}
\arguments{
\item{track}{Time format, either "99h99m99s" or "inf", see details}

\item{type}{Path and filename, file format must currently be ".txt"}
}
\value{
A ggplot2 object
}
\description{
track_mouse
}
\details{
visualize_mouse types
}
\examples{
visualize_mouse(mouse_track, "tinygrid")

}
