% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{calc_dist}
\alias{calc_dist}
\title{calc_dist}
\usage{
calc_dist(track, cumulative = TRUE)
}
\arguments{
\item{track}{Mouse track containing x any y coordinates}

\item{cumulative}{Return distance as cumulative sum or not, logical, default is true}
}
\value{
mouse track including distance
}
\description{
calc_dist
}
\details{
Calculate distance in pixels between track points that are of type 'Move'.
}
\examples{
calc_dist(mouse_df)

}
