% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devis-package.R
\docType{package}
\name{devis}
\alias{devis}
\alias{devis-package}
\title{devis}
\description{
Tracks and visualizes mouse movement data. blablabla
}
\author{
\strong{Maintainer}: Marco Sciaini \email{marco.sciaini@indblik.io}

Authors:
\itemize{
  \item Cedric Scherer
  \item Stefan Reifenberg
}

}
