% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_mouse.R
\name{track_mouse}
\alias{track_mouse}
\title{track_mouse}
\usage{
track_mouse(time, file, as_job = FALSE)
}
\arguments{
\item{time}{Time format, either "99h99m99s" or "inf", see details}

\item{file}{Path and filename, file format must currently be ".txt"}

\item{as_job}{Should the function be started as RStudio Job? (logical)}
}
\description{
track_mouse
}
\details{
track_mouse is a function that is best run as a background job, since it blocks
the R shell. See our README for a walkthrough of this.

The time format should be something of the following:
\itemize{
\item "3h23m04s"
\item "inf"
}

If you run
}
\examples{
\dontrun{
track_mouse("0h10m00s", "mouse_position.txt")
}
}
